Move getReferenceLabel to reference-data-utils
