+++ src/workspace-data/Data.js
@@ -45,8 +45,7 @@ import { EntityUploader } from './data-table/shared/EntityUploader';
 import WDSContent from './data-table/wds/WDSContent';
 import { WdsTroubleshooter } from './data-table/wds/WdsTroubleshooter';
 import { useImportJobs } from './import-jobs';
-import { getReferenceData } from './reference-data/reference-data-utils';
-import { getReferenceLabel } from './reference-data/reference-metadata';
+import { getReferenceData, getReferenceLabel } from './reference-data/reference-data-utils';
 import { ReferenceDataContent } from './reference-data/ReferenceDataContent';
 import { ReferenceDataDeleter } from './reference-data/ReferenceDataDeleter';
 import { ReferenceDataImporter } from './reference-data/ReferenceDataImporter';
